requirements.txt
Garmr==0.3
astroid==1.4.4
beautifulsoup4==4.4.0
colorama==0.3.7
gmpy2==2.0.7
lazy-object-proxy==1.2.1
nose==1.3.7
pycrypto==2.6.1
pylint==1.5.4
requests==2.7.0
six==1.10.0
vboxapi==1.0
virtualenv==15.0.0
wrapt==1.10.6
wsgiref==0.1.2
